+++ anime-flutter/anime_galaxy/lib/module_anime/state_manager/anime_details/anime_details.state_manager.dart
@@ -58,7 +58,7 @@ class AnimeDetailsStateManager {
 
   }
 
-  void rateAnime(int animeId,double rateValue){
+  void rateAnime(int animeId,int rateValue){
 
     this._animeDetailsService.rateAnime(animeId, rateValue).then((value) {
       if(value == null || value == false){
